modify label type in Datum from int to float

+++ include/caffe/data_layers.hpp
@@ -352,7 +352,7 @@ class ImageDataLayer : public BasePrefetchingDataLayer<Dtype> {
   virtual void ShuffleImages();
   virtual void InternalThreadEntry();
 
-  vector<std::pair<std::string, vector<int> > > lines_;
+  vector<std::pair<std::string, vector<float> > > lines_;
   int lines_id_;
 };
